requirements.txt
Pyyaml
tqdm==4.67.1
numpy==2.2.6
h5py==3.15.1
scikit-learn==1.7.2
torchinfo==1.8.0
matplotlib==3.10.7
python-dotenv==1.0.0
pytest==8.2
setuptools>=42
wheel
autopep8
